netusb: Refactor EEM code
Move EEM functionality from netusb to function_eem.
